er-13 has changed the type of notify_subscribers counter to int32_t from uint16_t which breaks builds of WITH_COAP=3 or 7.
This patch changes the 03 and 07 engine to int32_t.

An alternate fix would be to switch out the type in erbuim.h based on
WITH_COAP.

Or, maybe better, would be to drop old COAP versions.

+++ apps/er-coap-07/er-coap-07-observing.h
@@ -75,7 +75,7 @@ int coap_remove_observer_by_token(uip_ipaddr_t *addr, uint16_t port, uint8_t *to
 int coap_remove_observer_by_url(uip_ipaddr_t *addr, uint16_t port, const char *url);
 int coap_remove_observer_by_mid(uip_ipaddr_t *addr, uint16_t port, uint16_t mid);
 
-void coap_notify_observers(resource_t *resource, uint16_t obs_counter, void *notification);
+void coap_notify_observers(resource_t *resource, int32_t obs_counter, void *notification);
 
 void coap_observe_handler(resource_t *resource, void *request, void *response);
